Fix issue with googlenet module by using relative imports on Python 3.
This commit fixes an issue with importing
`sklearn_theano.feature_extraction` due to relative paths not being used
within that module.

The commit fixes the example on the sklean-theano homepage of recognising
cats and dogs in photo [1], clearly a primary use-case of any such machine
learning library.

Errors fixed (listed for searchability):

- No module named 'caffemodel'
- No module named 'googlenet_class_labels'
- No module named 'googlenet_layer_names'

[1] http://sklearn-theano.github.io/auto_examples/plot_multiple_localization.html